refactor(api): groups + invites → src/api modules (Effort 1)
Move the groups and invites data-access slices from
src/hooks/queries/groups (and its invites subfolder) to feature-sliced
src/api/groups and src/api/invites per ADR 0001. Each query endpoint
gains a per-file queryOptions factory; the scattered per-file group key
factories are consolidated into a single groupsKeys in
src/api/groups/types.ts, which also now owns the Group/GroupMember row
types (moved from src/types/groups.ts). Invites keep their own inviteKeys
factory. No behavior change; all consumers repointed.
