+++ src/api/groups/useDeleteGroup.ts
@@ -1,45 +1,49 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { userGroupsKeys } from "./useUserGroups";
+import { groupsKeys } from "./types";
 
 // Mutation function
-async function leaveGroup(variables: { groupId: string; userId: string }) {
+async function deleteGroup(variables: { groupId: string; userId: string }) {
   const { groupId, userId } = variables;
 
   const { error } = await supabase
-    .from("group_members")
-    .delete()
-    .eq("group_id", groupId)
-    .eq("user_id", userId);
+    .from("groups")
+    .update({ archived: true })
+    .eq("id", groupId)
+    .eq("created_by", userId);
 
   if (error) {
-    throw new Error("Failed to leave group");
+    throw new Error("Failed to delete group");
   }
 
   return true;
 }
 
 // Hook
-export function useLeaveGroupMutation() {
+export function useDeleteGroupMutation() {
   const queryClient = useQueryClient();
   const { toast } = useToast();
 
   return useMutation({
-    mutationFn: leaveGroup,
+    mutationFn: deleteGroup,
     onSuccess: (_data, variables) => {
+      // Invalidate all group-related queries
       queryClient.invalidateQueries({
-        queryKey: userGroupsKeys.user(variables.userId),
+        queryKey: groupsKeys.user(variables.userId),
+      });
+      queryClient.invalidateQueries({
+        queryKey: groupsKeys.all,
       });
       toast({
         title: "Success",
-        description: "Left group successfully",
+        description: "Group deleted successfully",
       });
     },
     onError: (error) => {
       toast({
         title: "Error",
-        description: error?.message || "Failed to leave group",
+        description: error?.message || "Failed to delete group",
         variant: "destructive",
       });
     },
