+++ src/api/groups/useGroupDetail.ts
@@ -1,13 +1,7 @@
-import { useQuery } from "@tanstack/react-query";
+import { queryOptions, useQuery } from "@tanstack/react-query";
 import { supabase } from "@/integrations/supabase/client";
-import type { Group } from "@/types/groups";
-
-// Query key factory
-export const groupDetailKeys = {
-  all: ["groups"] as const,
-  details: () => [...groupDetailKeys.all, "detail"] as const,
-  detail: (groupId: string) => [...groupDetailKeys.details(), groupId] as const,
-};
+import type { Group } from "./types";
+import { groupsKeys } from "./types";
 
 // Business logic function
 async function fetchGroupById(groupId: string): Promise<Group | null> {
@@ -25,11 +19,17 @@ async function fetchGroupById(groupId: string): Promise<Group | null> {
   return data;
 }
 
+export function groupDetailQuery(groupId: string) {
+  return queryOptions({
+    queryKey: groupsKeys.detail(groupId),
+    queryFn: () => fetchGroupById(groupId),
+  });
+}
+
 // Hook
 export function useGroupDetailQuery(groupId: string) {
   return useQuery({
-    queryKey: groupDetailKeys.detail(groupId),
-    queryFn: () => fetchGroupById(groupId),
+    ...groupDetailQuery(groupId),
     enabled: !!groupId,
   });
 }
