+++ src/api/groups/useInviteToGroup.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { groupMembersKeys } from "./useGroupMembers";
+import { groupsKeys } from "./types";
 
 // Helper functions
 async function findUserByUsernameOrEmail(usernameOrEmail: string) {
@@ -106,7 +106,7 @@ export function useInviteToGroupMutation(groupId: string) {
     onSuccess: (data) => {
       // Invalidate group members query to refresh the list
       queryClient.invalidateQueries({
-        queryKey: groupMembersKeys.members(groupId),
+        queryKey: groupsKeys.members(groupId),
       });
 
       toast({
